+++ src/openapi-renderer/helpers/cursorPaginationParamOpenapiProperty.ts
@@ -1,0 +1,6 @@
+export default function cursorPaginationParamOpenapiProperty() {
+  return {
+    type: ['string', 'null'],
+    description: 'Pagination cursor',
+  } as const
+}