+++ src/openapi-renderer/helpers/scrollPaginationCursorParamOpenapiProperty.ts
@@ -1,6 +1,0 @@
-export default function scrollPaginationCursorParamOpenapiProperty() {
-  return {
-    type: ['string', 'null'],
-    description: 'Scroll pagination cursor',
-  } as const
-}