+++ test-app/src/app/models/ApplicationModel.ts
@@ -1,10 +1,10 @@
 import { Dream } from '@rvoh/dream'
-import { DBClass } from '../../types/db.js'
+import { DB } from '../../types/db.js'
 import { connectionTypeConfig, schema } from '../../types/dream.js'
 import { globalTypeConfig } from '../../types/dream.globals.js'
 
 export default class ApplicationModel extends Dream {
-  public declare DB: DBClass
+  public declare DB: DB
 
   public override get schema() {
     return schema